move over getarticles to getter/service
